Fixed shader data precision mismatch error.

+++ source/engine/entity/GLGraphicPrimitive.cpp
@@ -6,39 +6,39 @@ void GLMesh::initialize()
 	glGenBuffers(1, &m_VBO);
 	glGenBuffers(1, &m_IBO);
 
-	std::vector<double> l_verticesBuffer;
+	std::vector<float> l_verticesBuffer;
 
 	std::for_each(m_vertices.begin(), m_vertices.end(), [&](Vertex val)
 	{
-		l_verticesBuffer.emplace_back(val.m_pos.x);
-		l_verticesBuffer.emplace_back(val.m_pos.y);
-		l_verticesBuffer.emplace_back(val.m_pos.z);
-		l_verticesBuffer.emplace_back(val.m_texCoord.x);
-		l_verticesBuffer.emplace_back(val.m_texCoord.y);
-		l_verticesBuffer.emplace_back(val.m_normal.x);
-		l_verticesBuffer.emplace_back(val.m_normal.y);
-		l_verticesBuffer.emplace_back(val.m_normal.z);
+		l_verticesBuffer.emplace_back((float)val.m_pos.x);
+		l_verticesBuffer.emplace_back((float)val.m_pos.y);
+		l_verticesBuffer.emplace_back((float)val.m_pos.z);
+		l_verticesBuffer.emplace_back((float)val.m_texCoord.x);
+		l_verticesBuffer.emplace_back((float)val.m_texCoord.y);
+		l_verticesBuffer.emplace_back((float)val.m_normal.x);
+		l_verticesBuffer.emplace_back((float)val.m_normal.y);
+		l_verticesBuffer.emplace_back((float)val.m_normal.z);
 	});
 
 	glBindVertexArray(m_VAO);
 
 	glBindBuffer(GL_ARRAY_BUFFER, m_VBO);
-	glBufferData(GL_ARRAY_BUFFER, l_verticesBuffer.size() * sizeof(double), &l_verticesBuffer[0], GL_STATIC_DRAW);
+	glBufferData(GL_ARRAY_BUFFER, l_verticesBuffer.size() * sizeof(float), &l_verticesBuffer[0], GL_STATIC_DRAW);
 
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_IBO);
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, m_indices.size() * sizeof(double), &m_indices[0], GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, m_indices.size() * sizeof(float), &m_indices[0], GL_STATIC_DRAW);
 
-	// position attribute, 1st attribution with 3 * sizeof(double) bits of data
+	// position attribute, 1st attribution with 3 * sizeof(float) bits of data
 	glEnableVertexAttribArray(0);
-	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 8 * sizeof(double), (void*)0);
+	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 8 * sizeof(float), (void*)0);
 
-	// texture attribute, 2nd attribution with 2 * sizeof(double) bits of data
+	// texture attribute, 2nd attribution with 2 * sizeof(float) bits of data
 	glEnableVertexAttribArray(1);
-	glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 8 * sizeof(double), (void*)(3 * sizeof(double)));
+	glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 8 * sizeof(float), (void*)(3 * sizeof(float)));
 
-	// normal coord attribute, 3rd attribution with 3 * sizeof(double) bits of data
+	// normal coord attribute, 3rd attribution with 3 * sizeof(float) bits of data
 	glEnableVertexAttribArray(2);
-	glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, 8 * sizeof(double), (void*)(5 * sizeof(double)));
+	glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, 8 * sizeof(float), (void*)(5 * sizeof(float)));
 
 	glBindBuffer(GL_ARRAY_BUFFER, 0);
 	glBindVertexArray(0);
@@ -363,11 +363,11 @@ void GLFrameBuffer::initialize()
 		glBindVertexArray(m_VAO);
 		glBindBuffer(GL_ARRAY_BUFFER, m_VBO);
 		// take care of std::vector's size and pointer of first element!!!
-		glBufferData(GL_ARRAY_BUFFER, m_Vertices.size() * sizeof(double), &m_Vertices[0], GL_STATIC_DRAW);
+		glBufferData(GL_ARRAY_BUFFER, m_Vertices.size() * sizeof(float), &m_Vertices[0], GL_STATIC_DRAW);
 		glEnableVertexAttribArray(0);
-		glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 5 * sizeof(double), (void*)0);
+		glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 5 * sizeof(float), (void*)0);
 		glEnableVertexAttribArray(1);
-		glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 5 * sizeof(double), (void*)(3 * sizeof(double)));
+		glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 5 * sizeof(float), (void*)(3 * sizeof(float)));
 		glBindBuffer(GL_ARRAY_BUFFER, 0);
 		glBindVertexArray(0);
 	}
